requirements.txt
Cython==0.29.21
einops==0.3.0
gluoncv==0.9.1
matplotlib==3.3.3
mxboard==0.1.0
mxnet-cu101==1.7.0
numpy==1.16.6
opencv-python==4.5.1.48
Pillow==8.1.0
pyaml==20.4.0
pycocotools-windows==2.0.0.2
pyparsing==2.4.7
PyYAML==5.3.1
scikit-image==0.18.1
scikit-learn==0.23.2
scikit-optimize==0.8.1
scipy==1.4.1
seaborn==0.11.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tqdm==4.56.0
wandb==0.10.31

